scala-job: 'sbt test' command does not pick up JVM options (#71)
The sbt 'test' command did not pick up the "javaOptions" option
from `build.sbt`.

The original problem was this was scoped down to "test" instead of
"Test". Update the sbt options to apply the "javaOptions" and "fork" to
all sbt commands.

Tested by running both "sbt test", "sbt run" and corresponding commands
in the sbt shell.