Enable Automated tagging workflow (#2361)
## Changes
Enable Automated tagging workflow. The workflow uses the `tagging.py`
script to:
* Flush the NEXT_CHANGELOG.md` file into the `CHANGELOG.md`.
* Create a Release commit and push it to `main` (without a PR)
* Create a tag for the release.

Usage and Changelog management details can be found in
[go/deco/release](go/deco/release).

This PR does:
* Added `cli_legacy` mode. Each SDK has it's own "mode" used to
determine which templates are to be used.
* `api_changelog: true` enables API changelog to be appended to
NEXT_CHANGELOG when generating the SDK.
* Removed `.codegen/changelog.md.tmpl`. Changelog templates are now part
of universe.
* Workflow and `tagging.py` files are synced from universe as part of
SDK generation. This allows us to test them and keep them in sync with
other SDKs.

## Tests
Workflow enabled in databricks/terraform-provider-databricks

---------

Co-authored-by: Andrew Nester <andrew.nester@databricks.com>