acc: Add CloudSlow config setting (#2470)
## Changes
New bool config setting CloudSlow in acceptance tests. If set, it
enables this test on Cloud but skips it in -short setting there. It does
not affect local runs, "Slow" is only applied to Cloud. The reason is
that we won't need to wait for cluster with mocked testserver.

Additionally, this setting enables -tail if -v is already enabled. 

## Why
Certain tests that use "bundle run" are too long due to starting a
cluster.

## Tests
Using this option in #2471