acc: Fix interactive_cluster test to use correct config; add NODE_TYPE_ID (#2478)
This is a follow up to #2471 where incorrect config was used in
interactive_cluster test.

## Changes
- Fixed interactive_cluster to use proper config, it was accidentally
referring to config from ../base
- Add $NODE_TYPE_ID env var and replacement to acceptance tests, this is
necessary for interactive_cluster test.
- Disable acceptance/bundle/override on cloud. This started failing
because it has real node type that gets replaced with NODE_TYPE_ID but
only in AWS env. Since the test is focussed on config merging, there is
no need to run it against real workspaces.
- Modify all tests in integration_whl to print rendered databricks.yml,
to prevent this kind of error.