Assert command exec id is correctly set (#2524)
## Changes
Modifies the telemetry acceptance test to assert that the same value of
`cmd-exec-id` is used in both the user agent and the telemetry payload.

To facilitate this, this PR also requires the acceptance testing
framework to write the request logs to disk without applying the
replacements.

## Why
Asserting a consistent value of `cmd-exec-id` is an important
correctness check because this is what we will use to get a completeness
signal for our telemetry tables.

## Tests
N/A

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->