Move `AccountClient` to `libs/command` (#2447)
## Changes
Move the AccountClient reader and setter from the root package to
`libs/command`.

## Why
This allows us to read the account client set in the context in all CLI
packages. This was not possible before because using
`root.AccountClient` would often result in an import cycle.

This is a natural follow-up to
https://github.com/databricks/cli/pull/2440 and
https://github.com/databricks/cli/pull/2444.

## Tests
Existing tests and one new unit test.

NO_CHANGELOG=true