Disable unit test in daemon package (#2538)
This unit test is inherently flaky and will fail if the gap between
process scheduling exceeds 10 seconds. I am hypothesizing that is what
happened in the failure here (or some equivalent issue, the exact cause
does not matter):

https://github.com/databricks/cli/actions/runs/13957485937/job/39072078260

Since we do not use the daemon package anywhere, let's disable this unit
test for now.