Add telemetry for `bundle init` (#2537)
## Changes
This PR adds telemetry logging for the `databricks bundle init` command.

## Why
We would like to know how often our templates are used and what revenues
are associated with each of our templates.

## Tests
Acceptance tests. Here are the properties that we want to ensure:
1. For Databricks templates, only enum/bool args should be captured;
others should be ignored. Non-enum args are not safe for centralization.
2. For user-defined custom templates, we should not log either the
template name or any of it's arguments (including boolean / enums).
3. The serialized bundle-uuid in `databricks.yml` should be the same as
the one recorded in the telemetry payload.


Also manually verified that the information shows up in our telemetry
table.