Minor refactoring of resources.go: rename a variable that collides with an imported package name; improve doc comments (#2544)
## Changes
1. Rename a local variable `resources` that collided with an imported
package of the same name
2. Changed doc comments for several methods to match the doc style of
other methods

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
More idiomatic source code is easier to read and comprehend

## Tests
<!-- How have you tested the changes? -->
Existing unit and integration tests are part of CI/CD checks

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->