Added `appproxy.Proxy` library (#2472)
## Changes
Added `appproxy.Proxy` library. This is a simple library which proxies
HTTP and Websocket request to the configured target URL and allows to
inject specified headers to all requests forwarded to the target. This
proxy is used for running Databricks apps locally.

It also adds 2 test components: `testserver.NewHttpServer` and
`testserver.NewWebsocketServer` which are used to start simple test
servers to be used in unit tests.

## Why
This library is needed later for a new CLI command which will allow to
locally run Databricks Apps. This proxy will be used to forward requests
to Databricks app which is running locally and will inject headers
needed for Databricks app (see
https://docs.databricks.com/aws/en/dev-tools/databricks-apps/app-development#what-http-headers-are-passed-to-databricks-apps)

## Tests
Added unit tests