[Python] Run acceptance tests against the latest snapshot (#2523)
## Changes
Run acceptance tests against the latest snapshot. We add matrix tests,
including both the latest snapshot and the latest release, so we don't
break compatibility.

We prebuild a wheel with the latest code and reuse it across acceptance
tests. The approach of installing from the source directory doesn't
work:
1. uv doesn't have a build tool, so it takes 10 seconds for setuptools
2. we can't reuse the same directory and cache, because setuptools can't
be used concurrently