Introduce new unit test for bindable resources (#2582)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
Add new unit test TestResourcesBindSupport to detect that all resources
meet preconditions for binding support

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
When a new resource is added, this unit test will fail if the resource
does not properly support `FindResourceByConfigKey` and `Exists`
methods. The author of a new resource would need to change these methods
to add the new resource, or explicitly set this resource as unbindable
in this unit test.

## Tests
<!-- How have you tested the changes? -->
New unit test is passing

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->