Convert default-python integration test to acceptance test (#2532)
## Changes
- Convert init_default_python_test.go to acceptance test
- Record dev targets in addition to prod targets.
- Record validate json output in addition to summary json.
- Add venv_activate test helper.

Note, GCP is disabled in this test due to differences from the aws and
azure (IsServicePrincipal is false there, causing different config).

Depends on #2535

## Why
init_default_python_test.go is one of a kind test - it uses golden
files, but it's a separate thing from the rest of acceptance tests. This
change makes it easier to update it with the rest of acceptance tests.