[Python] Add acceptance tests for locations (#2569)
## Changes
Add acceptance tests capturing how locations are recorded. 

Note that `--include-locations` doesn't include any location "deeper"
than "tasks," so we can't use it to test locations for properties with
"special" behavior like "notebook_path" that wouldn't have the same
location as other properties. For "notebook_path", we keep the location
of a virtual file in a bundle root.