Extract relative path visitors (#2588)
## Changes
Extract relative path visitors so we can implement path normalization
using them.

Replace `PathKind` with `TranslateMode` because there is no difference
in how we use them.

## Why
By splitting the logic that visits all fields that contain a path from
the code that performs the path translation, it becomes possible to make
this a 2-pass operation.

In the first pass, we make all paths relative to the bundle root
(incorporating their location). This can happen as early as possible.
Then, in the second pass, we know that all relative paths have already
been normalized to be relative to the bundle root, and we can perform
path translation without relying on the location information.

## Tests
Added unit tests in addition to the existing coverage