Use python script to extract command exec id (#2602)
## Why
The previous approach of using a pipeline was flaky on windows. Unix
tools have slightly different semantics on different operating systems
so it's hard to use and maintain them for acceptance tests.

That's why this PR switches to using a python script to extract the
command exec id.

## Tests
Existing tests pass.