Add formatting check to CI/CD (#2601)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
1. Add new steps to CI/CD workflow to check formatting issues
2. Introduce a new make target: fmtcheck, which allows to check
formatting without writing changes to disk
3. Run auto formatter on existing files to make the check pass

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

With the migration to golangci-lint v2.x, we lost automatic checks of
code formatters configured with it in `.golangci.yaml`. This PR restores
these checks.

## Tests
<!-- How have you tested the changes? -->
Lint step is passing

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->

---------

Co-authored-by: shreyas-goenka <88374338+shreyas-goenka@users.noreply.github.com>