replace $BUNDLE_NAME_SUFFIX with $UNIQUE_NAME in acceptance tests (#2628)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
Addresses a
[comment](https://github.com/databricks/cli/pull/2441#discussion_r1986833942)
from #2441 .
At the time of writing these two tests, UNIQUE_NAME variable did not
exist. This change makes the acceptance tests more consistent with other
tests on how they are adding a suffix to a bundle name

## Tests
<!-- How have you tested the changes? -->
Existing tests

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->