minor clean up: remove always-empty variables from log lines in acceptance_test.go (#2620)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
Addresses a
[comment](https://github.com/databricks/cli/pull/2516#discussion_r2011934945)
from #2516:
this change eliminates an unnecessary second call to `os.Getenv` when
the initial check has already confirmed the variable is empty

## Tests
<!-- How have you tested the changes? -->
Existing tests

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->