Add include/exclude flags to sync and bundle-sync commands (#2650)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
This allows users to specify file patterns to include / exclude from
syncs using respective flags in sync and bundle sync commands:

* `$ databricks sync <SRC> <DEST> --exclude "*.bak" --include
"node_modules/mylib"`
* `$ databricks bundle sync --exclude "*.bak" --include
"node_modules/mylib"`

[gitignore syntax](https://git-scm.com/docs/gitignore) for advanced
pattern matching is supported.

Note that in the current implementation excludes always takes higher
priority over includes. The order in which the flags are provided in the
prompt does not change this behavior. To overcome this behavior users
can provide an `--include` flag with a negated pattern (prefixed by a
`!`) instead of using an `--exclude` flag in the combinations of
patterns.

## Tests
<!-- How have you tested the changes? -->
Manual testing + introduced new acceptance tests

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->