Track deployment id in telemetry (#2648)
## Why
We need deployment_id to analyse distribution of bundle attributes like
number of resources, number of variables etc. Analysing them on the
bundle level is not possible because all bundles don't necessarily have
a bundle_uuid set.

Note: The server side protobuf changes for this are in flight.

## Tests
Acceptance test.

Also verified manually that the server returns a 200 status code with
the following response, even when an unknown field (`deployment_id`) in
this case is present:
```
< HTTP/2.0 200 OK
< {
<   "errors": null,
<   "numProtoSuccess": 1,
<   "numSuccess": 0
< } pid=43589 sdk=true
```