Remove pipeline 'deployment' from jsonschema (#2653)
## Changes
Remove pipeline `deployment` field from jsonschema

## Why
The `deployment` field is only meant to be set by DABs, and any
user-specified value is overridden. It's used to annotate which
pipelines are created by DABs. We have already removed `deployment` for
jobs and it was left for pipelines by mistake.

## Tests
Manually by inspecting jsonschema