acc: Do env var substitution in EnvMatrix values (#2678)
## Changes
When a $VAR reference is used in value of EnvMatrix, it'll be replaced
with its value if there is such variable previously defined.

## Why
It's useful to define variables in terms of other variables. For example
you can define "$MYCLI=$CLI --someparam". I'm using it to define CLI
command that either runs from bundle root or another directory.

## Tests
Unit tests + acceptance test.