[Feature] Migrate U2M Code to SDKs (#2076)
## Changes
Ports the SDK to this change:
https://github.com/databricks/databricks-sdk-go/pull/1108. This
eliminates the need for the SDK to invoke itself (!) when loading an
OAuth token. It also allows us to respond with more suitable errors when
fetching tokens. Functionally, this should behave identically to the
existing CLI from an external perspective.

One concrete improvement with this change is that the error message
returned when a refresh token is invalid now provides the correct
command to run, including the appropriate `--profile` flag.

Users will see the following message when using the CLI with a
profile/host with an expired refresh token:
![Screenshot 2025-01-20 at 13 18
59](https://github.com/user-attachments/assets/bf3040cb-971d-4c04-8557-0f6749d82b5e)


## Tests
Mostly this change involves refactoring existing code/tests to use/mock
the Go SDK interfaces, so there aren't too many substantial test
changes. `token_test.go` is rewritten using the table test approach to
reduce boilerplate.