dyn: replace Mapping.Set with Mapping.SetLoc (#2741)
## Changes
- Add Mapping.SetLoc(string, []Location, Value), which is a replacement
for Set() that does not return an error and only accepts string keys.
- Migrate all code to use this function.

## Why
The original function returned an error if key was holding value of
wrong type but the error was never checked. The new function does not
return an error.

This interface embraces string key rather than pretending that all types
are supported.

Note, in some cases this increases number of MustString() calls in the
code, however, this
- only happens where there already is MustString()
- is temporary - once we add a function to iterate over string keys
rather than values or pairs, it won't be necessary.

## Tests
Existing tests.