presets: Do not insert empty map for tags (#2756)
## Changes
- When applying presets, create map lazily, so that tags=nil remains
tags=nil and not updated to an empty map when no tags are defined in
presets.

## Why
- To avoid unnecessary changes to bundle.
- When calling API without terraform, `tags: {}` is end up being sent to
the backend. It's not the case with terraform, where empty tags are
dropped as this field is declared with `omitempty`
https://github.com/databricks/terraform-provider-databricks/blob/0964337/jobs/resource_job.go#L328

## Tests
Existing tests.

There is one acceptance test that incidentally tests the case when
user-provided tags are not empty:
acceptance/bundle/variables/resolve-empty