Improve error when target override fails (#2760)
## Changes
Add prefixes with context for errors in target override.

## Why
Saw an error that looks like
```
Error: invalid path '.'
```
The error was caused by some other breaking changes, so I don't have a
test, but still, if users ever see that it's not very helpful (e.g. path
does not refer to file system but it is related to dyn.Value).

## Tests
Existing tests.