Expand acceptance tests for empty resources (#2774)
## Changes
- Add new config setting EnvRepl which allows switching off replacements
for env vars. This allows switching replacements for things like '{}',
'null' which are frequent in JSON output and are easier to read if not
replaced.
- Refactor bundle/validate/empty_resources into subtests, one for each
type of content.
- Expand empty_resources tests with with_resources and with_grants.

## Why
Refactoring empty_resources into 3 subtests makes output.txt easier to
read.

Adding permissions and grants checks allows seeing what happens if
resource is partly initialized: outer struct is initialized but embedded
pointer is not.