Apply presets before applying defaults (#2778)
## Why
Presets should have more priority and need to see the configuration
before common defaults are applied, because this mutator is making
decisions based on presence of user-provided fields in the
configuration.

This does not matter currently where we only have few defaults, but will
start to matter once we add more, like in
https://github.com/databricks/cli/pull/2767

Previous PR https://github.com/databricks/cli/pull/2777 makes it
possible to move this without issues.

## Tests
Existing tests for empty resource added in
https://github.com/databricks/cli/pull/2771 and
https://github.com/databricks/cli/pull/2774