Add defaults on jobs that terraform applies (#2767)
## Changes
Initialize jobs resources with the defaults that terraform applies
(except for num_workers).

## Why
This makes those defaults visible for users (bundle validate -o json)
and PyDABs.

This makes terraform-removal branch simpler as there is no hidden
modification of resources that terraform does on deploy.

num_workers is skipped because it has a conflict with SingleNodeCluster
validator. We should find a better place for that (and other
validators), somewhere before applying defaults.