Read dashboard definition from `file_path` via filer interface (#2738)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
1. When dashboard resource contains a `file_path` field, that file is
read and its contents are set to the `serialized_dashboard` field. This
is done in a newly introduced `ConfigureDashboardSerializedDashboard`
mutator
2. Old behavior of wrapping file_path into terraform's `file` directive
is removed from `dashboardConverter` (unit tests for this are also
removed)
3. `ConfigureWSFS` mutator call is moved to be executed before
`PythonMutator` so that the latter has access to the Workspace File
System client

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
This change allows `bundle deploy` to correctly deploy dashboard
resource when running in Databricks Workspace / Runtime. Deployment does
not work currently, because reading a dashboard file via file system
interface throws an error in the Workspace.

## Tests
<!-- How have you tested the changes? -->
- Added a new dashboard deployment acceptance test
- Performed a manual test to deploy a dashboard from a workspace

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->