[Python] Handle deprecation and private annotations in codegen (#2773)
## Changes
Handle deprecated and private annotations in codegen.

- Deprecated fields are excluded from codegen
- Private fields are excluded from documentation

There is a trick to transitively mark fields as deprecated (private)
because it's normal for JSON schema to only mark fields as deprecated
(private) and not put the deprecated (private) annotation into
referenced schemas.

## Why
With that, we can continuously check that the Python code is consistent
with the latest JSON schema.