acc: Avoid printing too many unexpected files (#2795)
## Changes
If test detects an unexpected output file, use a heuristic on whether
it's useful to print its contents. Note, you can always see the contents
by running the test with -output setting, so it's okay to print less
there.

## Why
In some tests (e.g. experimental-jobs-as-code), where .venv is not
ignored but is located inside output folder, this prints a lot of
contents, making it hard to read the output.

## Tests
Manually caused experimental-jobs-as-code to see the effect.