Fix dynamic_version when sync root != bundle root (#2805)
## Changes
- When patching wheels with dynamic version, use paths relative to sync
root rather than dynamic root.
- Add debug output for cases where we don't do patching (was helpful
when investigating this issue). However, skip the whole thing if no
replacements are to be done (user is not using dynamic_version).

## Why
Previous implementation incorrectly assumed that those paths are
relative to bundle root, but we did not have a test with bundle root !=
sync root, so it was not caught. Inspired by
https://github.com/databricks/cli/issues/2784

## Tests
New test where wheel is referenced outside of bundle root and is patched
with dynamic version.