Upgrade to use Go SDK 0.68.0 (#2823)
## Changes
Upgrade to use Go SDK 0.68.0

It uses the same OpenAPI spec as 0.67.0 so no CLI commands changes

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

## Tests
<!-- How have you tested the changes? -->

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->