Sort file names in the "action" line output of sync commands (#2824)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

Addresses
[comment](https://github.com/databricks/cli/pull/2657#discussion_r2046683238)
in #2657: we want to get rid of non-determinism in output where
feasible. Deterministic output of CLI commands improves testing and any
possible automation that relies on the CLI.

## Tests
<!-- How have you tested the changes? -->
Existing acceptance tests were modified to not filter out the
non-deterministic line

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->