Add support for recording requests in integration acceptance tests (#2720)
## Changes
This PR adds support for recording API calls made during an integration
acceptance test by setting `RecordRequests = true` in the test.toml
file.

This PR adds a new struct `ProxyServer` which is a server that sits in
between a client and a real Databricks workspace. Having a server in
between allows us to record requests as they are sent to a real
Databricks workspace in our integration tests.

## Why
Other than being a useful feature in itself, this also unlocks migration
away from Terraform in DABs since we can now record the API requests
made by Terraform and ensure that our custom library makes the same API
requests in all cases.

## Tests
A bunch of new acceptance tests for the new feature.