Copy and sort lists of changed files when instantiating EventChanges (#2836)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

In #2824 in-place sorting was added which caused the slices to be
mutated on every event read and changed the order of the sync operations

## Tests
<!-- How have you tested the changes? -->

Existing acceptance test run with hyperfine

```
hyperfine -m 100 --show-output 'go test ./acceptance -run ^TestAccept$/^bundle$/^sync$/^dryrun$ -count=1'
```

was consistently failing on 20-30th iteration before this change and is
not failing on 100 iterations after the change.

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->

---------

Co-authored-by: Denis Bilenko <denis.bilenko@databricks.com>