fix acceptance tests for deploying and binding a dashboard (#2860)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
1. Change the golden files and the mocks to add the `pageType` field to
the serialized dashboard output

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
On May 13 2025 the Lakeview API started adding the `pageType`:
`PAGE_TYPE_UNSPECIFIED` to the serialized dashboard output

## Tests
<!-- How have you tested the changes? -->
Existing acceptance tests

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->