acc: disable Env replacements unless part of EnvMatrix (#2868)
## Changes
Default value of EnvRepl for a given key is no false unless the key is
part of EnvMatrix with multiple values.

## Why
Env variables that have values such as "0", "1", "yes" or "no" often
occur in text and usually not intended to be replaced.

Other Env variables, like Python version should be recorded unless we're
running the matrix and cannot have stable output without replacements.

## Tests
Existing tests.