Add ShellExecv method (#2862)
## Changes
ShellExecv is a variant of the Execv method that also configures a shell
to be used before running the command.

## Why
We need it for the scripts section in DABs:
https://github.com/databricks/cli/pull/2813.

Why:
1. We'll need to perform lexical analysis to figure out where the
boundaries for args in the script are. It's better to defer that to a
shell.
2. Direct calls to execv do not perform any environment variable
interpolation, so we need to go through a shell. This is neat because it
provides customers a way to configure inputs to their scripts.
3. Multiline scripts require a shell since, without them, we cannot
parse where a command ends.

Note: This uses the same priority order for shells as the
`artifacts.build` field, that is bash > sh > powershell.

## Tests
Unit test. Also acceptance tests in
https://github.com/databricks/cli/pull/2813