Render full diagnostics in `bundle run` (#2844)
## Why
`diags.Error()` only returns the`Details` in the diagnostic, and only
returns `Summary` if the details are empty. We should render both In
`bundle run` to be consistent with `bundle deploy`.

At the same time, this PR also renders all `warnings` and
`recommendations` if the user encounters an error. These warnings could
have helpful context on what they need to fix in their configuration.

## Tests
New acceptance test. Existing tests pass.