[Python] Fix missing support for floats (#2883)
## Changes
Add support for floats to Python code. Previously, attempting to
deserialize a float resulted in an error.

There is only one example of a float. It's in `spot_bid_max_price` in
`compute.AzureAttributes`, which was previously unsupported. Implement
support for floats with all implications of precision loss.

## Why
Fixes https://github.com/databricks/cli/issues/2875

## Tests
Unit tests and manually