Always use line feed for newlines in acceptance tests (#2865)
## Why
Based on feedback in
https://github.com/databricks/cli/pull/2720#discussion_r2084891783. We
should always treat new lines in acceptance tests in windows as `\n`.
This is important because the characters itself can be semantically
important in tests. For example, an additional `\r` character in windows
changes the upload payload if you are uploading the file's content to a
workspace.

## Tests
Existing workspace-io test should no longer need a custom override for
windows.