Remove the integration test TestDeployBasicBundleLogs (#2907)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

Remove a regression test introduced in #1711. This test aimed to capture
leaked logs to stderr. This test is obsolete with the new style of
acceptance tests, which compares the binary output with the golden file
- these existing tests will capture this kind of regression.

## Tests
Added an unconditional output line to stderr and confirmed that
acceptance tests capture it and fail the suite.