Convert TestBundleInitOnMlopsStacks into an acceptance test  (#2899)
## Changes
1. Convert TestBundleInitOnMlopsStacks into an acceptance test
2. Make `CLOUD_ENV_BASE` environment variable (with possible values of:
"azure", "aws", "gcp") available to the acceptance tests' script

## Why 
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

One change in the series of changes for converting integration tests
into acceptance tests.
This will allow for easier testing of various backing solutions for
bundle deployment