Fix acceptance tests for deploying and binding a dashboard (#2913)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
1. Change the golden files and the mocks to change the value of
`pageType` field to the serialized dashboard output

this is a follow-up change for #2860 

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
On May 20 2025 the Lakeview API changed the value of the `pageType`
field to `PAGE_TYPE_CANVAS` in the serialized dashboard output (when not
page type is specified)

## Tests
<!-- How have you tested the changes? -->
Existing acceptance tests

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->