Move deploy/dashboard acceptance test to a nested sub-folder (deploy/… (#2914)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

We will introduce more acceptance tests under `deploy/dashboard` in the
follow-up PRs

## Tests
<!-- How have you tested the changes? -->
Existing test is unchanged and must pass

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->