Expand secret scope permissions test with more examples (#2890)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
This test now exercises the top-level permissions mapping and setting
secret scopes permissions on the resource level

## Tests
<!-- How have you tested the changes? -->
Extended existing test

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->