Convert TestBundleDestroy to acceptance test  (#2912)
## Changes
Additionally, this changes `title` implementation in `script.prepare` to
print new lines if present

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

One change in the series of changes for converting integration tests
into acceptance tests.
This will allow for easier testing of various backing solutions for
bundle deployment