testserver: log error on "method not allowed" same as "not found" (#2933)
This extends the diagnostics we have for 404 to 405 errors.

Today, if you call testserver and there is no path registered, you'll
get 404 and a nice error message suggesting which setting to add to
config. If you call some path that is present but does not have handler
for the required method, you'll get 405 silently and without errors.

The deploy-artifact-path-type test had to be fixed since it relied on
undefined handler before and now it fails.