Translate relative paths `root_path` and `glob.include` for pipelines (#2931)
## Why
Translating these fields is necessary for Lakeflow authoring experience
to work well with DABs in the workspace. We expect users to set these
fields manually in their bundle config.

For:
1. `root_path`: We translate the relative root_path to an absolute path
in the workspace. Since we assume that the root_path is within the sync
root for the bundle (i.e., it's managed by the bundle), we error if it
does not exist or lies outside the bundle root.
2. `glob.include`: We directly translate relative glob patterns to
absolute ones. Performing any client-side validation requires keeping
the validation logic in sync with the server-side definition, so for
now, we'll omit performing any client-side validation for glob patterns
specified in `glob.include`.

## Tests
Acceptance tests

---------

Co-authored-by: Pieter Noordhuis <pieter.noordhuis@databricks.com>